requirements.txt
pytorch-lightning>=1.9,<2.0
hydra-core>=1.3,<1.4
hydra_colorlog
wandb
rich
transformers>=4.26,<4.27
black[d]
